Backend_DockLaunch/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
GitPython==3.1.42
docker==7.0.0
requests==2.31.0
python-dotenv==1.0.1